src/screen: relocate inline tests to screen_tests.rs
screen.rs reaches 1,465 LOC with a 941-LOC `#[cfg(test)] mod tests` block.
Same playbook as #1080-#1089: extract to a sibling screen_tests.rs.

After: screen.rs drops from 1,465 → 527 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
